+++ example/lib/dialog/message_dialog.dart
@@ -2,7 +2,6 @@ import 'dart:io';
 
 import 'package:flutter/material.dart';
 import 'package:nav_hooks/dialog/dialog.dart';
-import 'package:nav_hooks/dialog/dialog_state.dart';
 import 'package:nav_hooks/enum/enum_nav_ani.dart';
 
 class MessageDialog extends DialogWidget {